pyproject.toml
[tool.poetry]
name = "snowball"
version = "0.1.0"
description = ""
authors = ["Elven Lin, RetrO9 <elvenlin1998@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.4.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
